4.4.1.57 (2021-05-12) 更新 更新为Mirai 2.6.4 新增 人工智障回复 新增 防刷屏 修复 获取时间戳的单位问题 修复 撤回消息发送器的参数问题

+++ src/main/java/ninja/skyrocketing/fuyao/bot/sender/group/GroupMessageSender.java
@@ -122,7 +122,7 @@ public class GroupMessageSender {
      * @param group Group
      * @param recall Integer
      */
-    public static void sendMessageByGroupId(MessageChainBuilder message, Group group, Integer recall) throws IOException {
+    public static void sendMessageByGroupId(MessageChainBuilder message, Group group, Long recall) throws IOException {
         group.sendMessage(message.asMessageChain()).recallIn(recall);
         LogUtil.messageLog("(" + recall + " 毫秒后撤回) " + message.toString().replaceAll("\n", ""),
                 group.getId(),
